+++ backend/app/api/v1/endpoints/likes.py
@@ -4,29 +4,33 @@ from fastapi import APIRouter, HTTPException
 
 router = APIRouter()
 
-@router.get("likes/posts/{id}")
-def getLikesForPostById(id):
+
+@router.get("/posts/{id}")
+def get_likes_by_post_id(id):
     try:
         return {}
     except HTTPException:
         pass 
 
-@router.get("likes/comments/{id}")
-def getLikesForCommentById(id):
+
+@router.get("/comments/{id}")
+def get_likes_by_comment_id(id):
     try:
         return {}
     except HTTPException:
         pass 
 
-@router.post("likes/comments")
-def likeComment():
+
+@router.post("/comments")
+def manage_like_on_comment():
     try:
         return {}
     except HTTPException:
         pass 
 
-@router.post("likes/posts")
-def likePost():
+
+@router.post("/posts")
+def manage_like_on_post():
     try:
         return {}
     except HTTPException: